materialize-postgres: replace with materialize-postgres-rc
This is a mechanical change, deleting the existing materialize-postgres and replacing it with
materialize-postgres-rc.
