--- materialize-postgres-rc/.snapshots/TestSpecification
@@ -1,102 +1,0 @@
-{
-  "endpoint_spec_schema_json": {
-    "$schema": "http://json-schema.org/draft/2020-12/schema",
-    "$id": "https://github.com/estuary/connectors/materialize-postgres-rc/config",
-    "properties": {
-      "address": {
-        "type": "string",
-        "title": "Address",
-        "description": "Host and port of the database."
-      },
-      "user": {
-        "type": "string",
-        "title": "User",
-        "description": "Database user to connect as."
-      },
-      "password": {
-        "type": "string",
-        "title": "Password",
-        "description": "Password for the specified database user.",
-        "secret": true
-      },
-      "database": {
-        "type": "string",
-        "title": "Database",
-        "description": "Name of the logical database to materialize to."
-      },
-      "schema": {
-        "type": "string",
-        "title": "Database Schema",
-        "description": "Database schema for bound collection tables (unless overridden within the binding resource configuration) as well as associated materialization metadata tables",
-        "default": "public"
-      },
-      "networkTunnel": {
-        "properties": {
-          "sshForwarding": {
-            "properties": {
-              "sshEndpoint": {
-                "type": "string",
-                "title": "SSH Endpoint",
-                "description": "Endpoint of the remote SSH server that supports tunneling (in the form of ssh://user@hostname[:port])",
-                "pattern": "^ssh://.+@.+$"
-              },
-              "privateKey": {
-                "type": "string",
-                "title": "SSH Private Key",
-                "description": "Private key to connect to the remote SSH server.",
-                "multiline": true,
-                "secret": true
-              }
-            },
-            "additionalProperties": false,
-            "type": "object",
-            "required": [
-              "sshEndpoint",
-              "privateKey"
-            ],
-            "title": "SSH Forwarding"
-          }
-        },
-        "additionalProperties": false,
-        "type": "object",
-        "title": "Network Tunnel",
-        "description": "Connect to your system through an SSH server that acts as a bastion host for your network."
-      }
-    },
-    "type": "object",
-    "required": [
-      "address",
-      "user",
-      "password"
-    ],
-    "title": "SQL Connection"
-  },
-  "resource_spec_schema_json": {
-    "$schema": "http://json-schema.org/draft/2020-12/schema",
-    "$id": "https://github.com/estuary/connectors/materialize-postgres-rc/table-config",
-    "properties": {
-      "schema": {
-        "type": "string",
-        "title": "Alternative Schema",
-        "description": "Alternative schema for this table (optional)"
-      },
-      "table": {
-        "type": "string",
-        "title": "Table",
-        "description": "Name of the database table"
-      },
-      "delta_updates": {
-        "type": "boolean",
-        "title": "Delta Update",
-        "description": "Should updates to this table be done via delta updates. Defaults is false.",
-        "default": false
-      }
-    },
-    "type": "object",
-    "required": [
-      "table"
-    ],
-    "title": "SQL Table"
-  },
-  "documentation_url": "https://go.estuary.dev/materialize-postgresql"
-}
