--- materialize-postgres-rc/CHANGELOG.md
@@ -1,5 +1,0 @@
-# materialize-postgres-rc
-
-## v1, 2022-07-27
-
-- Beginning of changelog.
