--- materialize-postgres-rc/VERSION
@@ -1,1 +1,0 @@
-v1
