--- materialize-postgres-rc/reserved_words.go
@@ -1,479 +1,0 @@
-package main
-
-// This list was obtained by filtering all fields in this table [0]
-// that were marked as being reserved in any version of SQL or PostgreSQL.
-// I used this AWK script to obtain it after copy-pasting the table in a text file:
-// awk '/\treserved/ { print $1 }' postgres-table > reserved
-// [0]: https://www.postgresql.org/docs/current/sql-keywords-appendix.html
-var PG_RESERVED_WORDS = []string{
-	"abs",
-	"absolute",
-	"acos",
-	"action",
-	"add",
-	"all",
-	"allocate",
-	"alter",
-	"analyse",
-	"analyze",
-	"and",
-	"any",
-	"are",
-	"array",
-	"array_agg",
-	"array_max_cardinality",
-	"as",
-	"asc",
-	"asensitive",
-	"asin",
-	"assertion",
-	"asymmetric",
-	"at",
-	"atan",
-	"atomic",
-	"authorization",
-	"avg",
-	"begin",
-	"begin_frame",
-	"begin_partition",
-	"between",
-	"bigint",
-	"binary",
-	"bit",
-	"bit_length",
-	"blob",
-	"boolean",
-	"both",
-	"by",
-	"call",
-	"called",
-	"cardinality",
-	"cascade",
-	"cascaded",
-	"case",
-	"cast",
-	"catalog",
-	"ceil",
-	"ceiling",
-	"char",
-	"character",
-	"character_length",
-	"char_length",
-	"check",
-	"classifier",
-	"clob",
-	"close",
-	"coalesce",
-	"collate",
-	"collation",
-	"collect",
-	"column",
-	"commit",
-	"concurrently",
-	"condition",
-	"connect",
-	"connection",
-	"constraint",
-	"constraints",
-	"contains",
-	"continue",
-	"convert",
-	"copy",
-	"corr",
-	"corresponding",
-	"cos",
-	"cosh",
-	"count",
-	"covar_pop",
-	"covar_samp",
-	"create",
-	"cross",
-	"cube",
-	"cume_dist",
-	"current",
-	"current_catalog",
-	"current_date",
-	"current_default_transform_group",
-	"current_path",
-	"current_role",
-	"current_row",
-	"current_schema",
-	"current_time",
-	"current_timestamp",
-	"current_transform_group_for_type",
-	"current_user",
-	"cursor",
-	"cycle",
-	"datalink",
-	"date",
-	"day",
-	"deallocate",
-	"dec",
-	"decfloat",
-	"decimal",
-	"declare",
-	"default",
-	"deferrable",
-	"deferred",
-	"define",
-	"delete",
-	"dense_rank",
-	"deref",
-	"desc",
-	"describe",
-	"descriptor",
-	"deterministic",
-	"diagnostics",
-	"disconnect",
-	"distinct",
-	"dlnewcopy",
-	"dlpreviouscopy",
-	"dlurlcomplete",
-	"dlurlcompleteonly",
-	"dlurlcompletewrite",
-	"dlurlpath",
-	"dlurlpathonly",
-	"dlurlpathwrite",
-	"dlurlscheme",
-	"dlurlserver",
-	"dlvalue",
-	"do",
-	"domain",
-	"double",
-	"drop",
-	"dynamic",
-	"each",
-	"element",
-	"else",
-	"empty",
-	"end",
-	"end-exec",
-	"end_frame",
-	"end_partition",
-	"equals",
-	"escape",
-	"every",
-	"except",
-	"exception",
-	"exec",
-	"execute",
-	"exists",
-	"exp",
-	"external",
-	"extract",
-	"false",
-	"fetch",
-	"filter",
-	"first",
-	"first_value",
-	"float",
-	"floor",
-	"for",
-	"foreign",
-	"found",
-	"frame_row",
-	"free",
-	"freeze",
-	"from",
-	"full",
-	"function",
-	"fusion",
-	"get",
-	"global",
-	"go",
-	"goto",
-	"grant",
-	"group",
-	"grouping",
-	"groups",
-	"having",
-	"hold",
-	"hour",
-	"identity",
-	"ilike",
-	"immediate",
-	"import",
-	"in",
-	"indicator",
-	"initial",
-	"initially",
-	"inner",
-	"inout",
-	"input",
-	"insensitive",
-	"insert",
-	"int",
-	"integer",
-	"intersect",
-	"intersection",
-	"interval",
-	"into",
-	"is",
-	"isnull",
-	"isolation",
-	"join",
-	"json_array",
-	"json_arrayagg",
-	"json_exists",
-	"json_object",
-	"json_objectagg",
-	"json_query",
-	"json_table",
-	"json_table_primitive",
-	"json_value",
-	"key",
-	"lag",
-	"language",
-	"large",
-	"last",
-	"last_value",
-	"lateral",
-	"lead",
-	"leading",
-	"left",
-	"level",
-	"like",
-	"like_regex",
-	"limit",
-	"listagg",
-	"ln",
-	"local",
-	"localtime",
-	"localtimestamp",
-	"log",
-	"log10",
-	"lower",
-	"match",
-	"matches",
-	"match_number",
-	"match_recognize",
-	"max",
-	"measures",
-	"member",
-	"merge",
-	"method",
-	"min",
-	"minute",
-	"mod",
-	"modifies",
-	"module",
-	"month",
-	"multiset",
-	"names",
-	"national",
-	"natural",
-	"nchar",
-	"nclob",
-	"new",
-	"next",
-	"no",
-	"none",
-	"normalize",
-	"not",
-	"notnull",
-	"nth_value",
-	"ntile",
-	"null",
-	"nullif",
-	"numeric",
-	"occurrences_regex",
-	"octet_length",
-	"of",
-	"offset",
-	"old",
-	"omit",
-	"on",
-	"one",
-	"only",
-	"open",
-	"option",
-	"or",
-	"order",
-	"out",
-	"outer",
-	"output",
-	"over",
-	"overlaps",
-	"overlay",
-	"pad",
-	"parameter",
-	"partial",
-	"partition",
-	"pattern",
-	"per",
-	"percent",
-	"percentile_cont",
-	"percentile_disc",
-	"percent_rank",
-	"period",
-	"permute",
-	"placing",
-	"portion",
-	"position",
-	"position_regex",
-	"power",
-	"precedes",
-	"precision",
-	"prepare",
-	"preserve",
-	"primary",
-	"prior",
-	"privileges",
-	"procedure",
-	"ptf",
-	"public",
-	"range",
-	"rank",
-	"read",
-	"reads",
-	"real",
-	"recursive",
-	"ref",
-	"references",
-	"referencing",
-	"regr_avgx",
-	"regr_avgy",
-	"regr_count",
-	"regr_intercept",
-	"regr_r2",
-	"regr_slope",
-	"regr_sxx",
-	"regr_sxy",
-	"regr_syy",
-	"relative",
-	"release",
-	"restrict",
-	"result",
-	"return",
-	"returning",
-	"returns",
-	"revoke",
-	"right",
-	"rollback",
-	"rollup",
-	"row",
-	"rows",
-	"row_number",
-	"running",
-	"savepoint",
-	"schema",
-	"scope",
-	"scroll",
-	"search",
-	"second",
-	"section",
-	"seek",
-	"select",
-	"sensitive",
-	"session",
-	"session_user",
-	"set",
-	"show",
-	"similar",
-	"sin",
-	"sinh",
-	"size",
-	"skip",
-	"smallint",
-	"some",
-	"space",
-	"specific",
-	"specifictype",
-	"sql",
-	"sqlcode",
-	"sqlerror",
-	"sqlexception",
-	"sqlstate",
-	"sqlwarning",
-	"sqrt",
-	"start",
-	"static",
-	"stddev_pop",
-	"stddev_samp",
-	"submultiset",
-	"subset",
-	"substring",
-	"substring_regex",
-	"succeeds",
-	"sum",
-	"symmetric",
-	"system",
-	"system_time",
-	"system_user",
-	"table",
-	"tablesample",
-	"tan",
-	"tanh",
-	"temporary",
-	"then",
-	"time",
-	"timestamp",
-	"timezone_hour",
-	"timezone_minute",
-	"to",
-	"trailing",
-	"transaction",
-	"translate",
-	"translate_regex",
-	"translation",
-	"treat",
-	"trigger",
-	"trim",
-	"trim_array",
-	"true",
-	"truncate",
-	"uescape",
-	"union",
-	"unique",
-	"unknown",
-	"unmatched",
-	"unnest",
-	"update",
-	"upper",
-	"usage",
-	"user",
-	"using",
-	"value",
-	"values",
-	"value_of",
-	"varbinary",
-	"varchar",
-	"variadic",
-	"varying",
-	"var_pop",
-	"var_samp",
-	"verbose",
-	"versioning",
-	"view",
-	"when",
-	"whenever",
-	"where",
-	"width_bucket",
-	"window",
-	"with",
-	"within",
-	"without",
-	"work",
-	"write",
-	"xml",
-	"xmlagg",
-	"xmlattributes",
-	"xmlbinary",
-	"xmlcast",
-	"xmlcomment",
-	"xmlconcat",
-	"xmldocument",
-	"xmlelement",
-	"xmlexists",
-	"xmlforest",
-	"xmliterate",
-	"xmlnamespaces",
-	"xmlparse",
-	"xmlpi",
-	"xmlquery",
-	"xmlserialize",
-	"xmltable",
-	"xmltext",
-	"xmlvalidate",
-	"year",
-}
