+++ materialize-postgres/.snapshots/TestFencingCases-straddle
