+++ materialize-postgres/.snapshots/TestSpecification
@@ -28,6 +28,13 @@
         "description": "Name of the logical database to materialize to.",
         "order": 3
       },
+      "schema": {
+        "type": "string",
+        "title": "Database Schema",
+        "description": "Database schema for bound collection tables (unless overridden within the binding resource configuration) as well as associated materialization metadata tables",
+        "default": "public",
+        "order": 4
+      },
       "networkTunnel": {
         "properties": {
           "sshForwarding": {
@@ -73,8 +80,21 @@
     "$schema": "http://json-schema.org/draft/2020-12/schema",
     "$id": "https://github.com/estuary/connectors/materialize-postgres/table-config",
     "properties": {
+      "schema": {
+        "type": "string",
+        "title": "Alternative Schema",
+        "description": "Alternative schema for this table (optional)"
+      },
       "table": {
-        "type": "string"
+        "type": "string",
+        "title": "Table",
+        "description": "Name of the database table"
+      },
+      "delta_updates": {
+        "type": "boolean",
+        "title": "Delta Update",
+        "description": "Should updates to this table be done via delta updates. Defaults is false.",
+        "default": false
       }
     },
     "type": "object",
