+++ materialize-postgres/Dockerfile
@@ -15,13 +15,15 @@ RUN go mod download
 RUN curl -L --proto '=https' --tlsv1.2 -sSf "https://github.com/estuary/flow/releases/download/dev/flow-x86-linux.tar.gz"| tar -zx -C /usr/local/bin/
 
 # Build the connector projects we depend on.
+COPY go-schema-gen           ./go-schema-gen
 COPY materialize-boilerplate ./materialize-boilerplate
 COPY materialize-postgres    ./materialize-postgres
+COPY materialize-sql         ./materialize-sql
 COPY testsupport             ./testsupport
 COPY go-network-tunnel       ./go-network-tunnel
 
 # Test and build the connector.
-RUN go test  -tags nozstd -v ./materialize-postgres/...
+RUN go test  -tags nozstd,nodb -v ./materialize-postgres/...
 RUN go build -tags nozstd -v -o ./connector ./materialize-postgres/...
 
 # Runtime Stage
