+++ materialize-postgres/driver.go
@@ -2,21 +2,23 @@ package main
 
 import (
 	"context"
-	"database/sql"
+	stdsql "database/sql"
 	"encoding/json"
 	"fmt"
 	"net"
 	"net/url"
 	"strings"
+	"text/template"
+	"time"
 
 	networkTunnel "github.com/estuary/connectors/go-network-tunnel"
 	boilerplate "github.com/estuary/connectors/materialize-boilerplate"
+	sql "github.com/estuary/connectors/materialize-sql"
 	pf "github.com/estuary/flow/go/protocols/flow"
 	pm "github.com/estuary/flow/go/protocols/materialize"
-	sqlDriver "github.com/estuary/flow/go/protocols/materialize/sql"
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgx/v4"
-	pgxStd "github.com/jackc/pgx/v4/stdlib"
+	_ "github.com/jackc/pgx/v4/stdlib"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -30,12 +32,12 @@ type tunnelConfig struct {
 }
 
 // config represents the endpoint configuration for postgres.
-// It must match the one defined for the source specs (flow.yaml) in Rust.
 type config struct {
 	Address  string `json:"address" jsonschema:"title=Address,description=Host and port of the database." jsonschema_extras:"order=0"`
 	User     string `json:"user" jsonschema:"title=User,description=Database user to connect as." jsonschema_extras:"order=1"`
 	Password string `json:"password" jsonschema:"title=Password,description=Password for the specified database user." jsonschema_extras:"secret=true,order=2"`
 	Database string `json:"database,omitempty" jsonschema:"title=Database,description=Name of the logical database to materialize to." jsonschema_extras:"order=3"`
+	Schema   string `json:"schema,omitempty" jsonschema:"title=Database Schema,default=public,description=Database schema for bound collection tables (unless overridden within the binding resource configuration) as well as associated materialization metadata tables" jsonschema_extras:"order=4"`
 
 	NetworkTunnel *tunnelConfig `json:"networkTunnel,omitempty" jsonschema:"title=Network Tunnel,description=Connect to your system through an SSH server that acts as a bastion host for your network."`
 }
@@ -71,11 +73,24 @@ func (c *config) ToURI() string {
 	if c.Database != "" {
 		uri.Path = "/" + c.Database
 	}
+	// Compatibility with pgBouncer. See:
+	// https://github.com/jackc/pgx/issues/650
+	uri.RawQuery = "statement_cache_mode=describe"
 	return uri.String()
 }
 
 type tableConfig struct {
-	Table string `json:"table"`
+	Schema string `json:"schema,omitempty" jsonschema:"title=Alternative Schema,description=Alternative schema for this table (optional)"`
+	Table  string `json:"table" jsonschema:"title=Table,description=Name of the database table"`
+	Delta  bool   `json:"delta_updates,omitempty" jsonschema:"default=false,title=Delta Update,description=Should updates to this table be done via delta updates. Defaults is false."`
+}
+
+func newTableConfig(ep *sql.Endpoint) sql.Resource {
+	return &tableConfig{
+		// Default to an explicit endpoint configuration schema, if set.
+		// This will be over-written by a present `schema` property within `raw`.
+		Schema: ep.Config.(*config).Schema,
+	}
 }
 
 // Validate the resource configuration.
@@ -86,24 +101,26 @@ func (r tableConfig) Validate() error {
 	return nil
 }
 
-func (c tableConfig) Path() sqlDriver.ResourcePath {
+func (c tableConfig) Path() sql.TablePath {
+	if c.Schema != "" {
+		return []string{c.Schema, c.Table}
+	}
 	return []string{c.Table}
 }
 
 func (c tableConfig) DeltaUpdates() bool {
-	return false // PostgreSQL doesn't support delta updates.
+	return c.Delta
 }
 
 func newPostgresDriver() pm.DriverServer {
-	return &sqlDriver.Driver{
+	return &sql.Driver{
 		DocumentationURL: "https://go.estuary.dev/materialize-postgresql",
 		EndpointSpecType: new(config),
 		ResourceSpecType: new(tableConfig),
-		NewResource:      func(sqlDriver.Endpoint) sqlDriver.Resource { return new(tableConfig) },
-		NewEndpoint: func(ctx context.Context, raw json.RawMessage) (sqlDriver.Endpoint, error) {
+		NewEndpoint: func(ctx context.Context, raw json.RawMessage) (*sql.Endpoint, error) {
 			var cfg = new(config)
 			if err := pf.UnmarshalStrict(raw, cfg); err != nil {
-				return nil, fmt.Errorf("parsing Postgresql configuration: %w", err)
+				return nil, fmt.Errorf("parsing endpoint configuration: %w", err)
 			}
 
 			log.WithFields(log.Fields{
@@ -111,6 +128,12 @@ func newPostgresDriver() pm.DriverServer {
 				"address":  cfg.Address,
 				"user":     cfg.User,
 			}).Info("opening database")
+
+			var metaBase sql.TablePath
+			if cfg.Schema != "" {
+				metaBase = append(metaBase, cfg.Schema)
+			}
+			var metaSpecs, metaCheckpoints = sql.MetaTables(metaBase)
 
 			// If SSH Endpoint is configured, then try to start a tunnel before establishing connections
 			if cfg.NetworkTunnel != nil && cfg.NetworkTunnel.SshForwarding != nil && cfg.NetworkTunnel.SshForwarding.SshEndpoint != "" {
@@ -137,315 +160,259 @@ func newPostgresDriver() pm.DriverServer {
 				}
 			}
 
-			db, err := sql.Open("pgx", cfg.ToURI())
-			if err != nil {
-				return nil, fmt.Errorf("opening Postgres database: %w", err)
-			}
-			return sqlDriver.NewStdEndpoint(cfg, db, PostgresSQLGenerator(), sqlDriver.DefaultFlowTables("")), nil
-		},
-		NewTransactor: func(
-			ctx context.Context,
-			epi sqlDriver.Endpoint,
-			spec *pf.MaterializationSpec,
-			fence sqlDriver.Fence,
-			resources []sqlDriver.Resource,
-		) (_ pm.Transactor, err error) {
-			var ep = epi.(*sqlDriver.StdEndpoint)
-			var d = &transactor{
-				gen: ep.Generator(),
-			}
-			d.store.fence = fence.(*sqlDriver.StdFence)
-
-			// Establish connections.
-			if d.load.conn, err = pgxStd.AcquireConn(ep.DB()); err != nil {
-				return nil, fmt.Errorf("load pgx.AcquireConn: %w", err)
-			}
-			if d.store.conn, err = pgxStd.AcquireConn(ep.DB()); err != nil {
-				return nil, fmt.Errorf("store pgx.AcquireConn: %w", err)
-			}
-
-			for _, spec := range spec.Bindings {
-				var target = sqlDriver.ResourcePath(spec.ResourcePath).Join()
-				if err = d.addBinding(ctx, spec); err != nil {
-					return nil, fmt.Errorf("%s: %w", target, err)
-				}
-			}
-
-			// Build a query which unions the results of each load subquery.
-			var subqueries []string
-			for _, b := range d.bindings {
-				subqueries = append(subqueries, b.load.query.sql)
-			}
-			var loadAllSQL = strings.Join(subqueries, "\nUNION ALL\n") + ";"
-
-			d.load.stmt, err = d.load.conn.Prepare(ctx, "load-join-all", loadAllSQL)
-			if err != nil {
-				return nil, fmt.Errorf("conn.PrepareContext(%s): %w", loadAllSQL, err)
-			}
-
-			return d, nil
+			return &sql.Endpoint{
+				Config:              cfg,
+				Dialect:             pgDialect,
+				MetaSpecs:           metaSpecs,
+				MetaCheckpoints:     &metaCheckpoints,
+				Client:              client{uri: cfg.ToURI()},
+				CreateTableTemplate: tplCreateTargetTable,
+				NewResource:         newTableConfig,
+				NewTransactor:       newTransactor,
+			}, nil
 		},
 	}
 }
 
-type transactor struct {
-	gen *sqlDriver.Generator
+// client implements the sql.Client interface.
+type client struct {
+	uri string
+}
 
+func (c client) FetchSpecAndVersion(ctx context.Context, specs sql.Table, materialization pf.Materialization) (specB64, version string, err error) {
+	err = c.withDB(func(db *stdsql.DB) error {
+		specB64, version, err = sql.StdFetchSpecAndVersion(ctx, db, specs, materialization)
+		return err
+	})
+	return
+}
+
+func (c client) ExecStatements(ctx context.Context, statements []string) error {
+	return c.withDB(func(db *stdsql.DB) error { return sql.StdSQLExecStatements(ctx, db, statements) })
+}
+
+func (c client) InstallFence(ctx context.Context, checkpoints sql.Table, fence sql.Fence) (sql.Fence, error) {
+	var err = c.withDB(func(db *stdsql.DB) error {
+		// Pick a path randomly to ensure both are exercised and correct.
+		if time.Now().Unix()%2 == 0 {
+
+			// Option 1: Install using template.
+			var query strings.Builder
+			if err := tplInstallFence.Execute(&query, fence); err != nil {
+				return fmt.Errorf("evaluating fence template: %w", err)
+			}
+			return db.QueryRow(query.String()).Scan(&fence.Fence, &fence.Checkpoint)
+		}
+
+		// Option 2: Install using StdInstallFence.
+		var err error
+		fence, err = sql.StdInstallFence(ctx, db, checkpoints, fence)
+		return err
+	})
+	return fence, err
+}
+
+func (c client) withDB(fn func(*stdsql.DB) error) error {
+	var db, err = stdsql.Open("pgx", c.uri)
+	if err != nil {
+		return err
+	}
+	defer db.Close()
+	return fn(db)
+}
+
+type transactor struct {
 	// Variables exclusively used by Load.
 	load struct {
-		conn *pgx.Conn
-		stmt *pgconn.StatementDescription
+		conn     *pgx.Conn
+		unionSQL string
 	}
 	// Variables accessed by Prepare, Store, and Commit.
 	store struct {
-		batch *pgx.Batch
+		batch pgBatch
 		conn  *pgx.Conn
-		fence *sqlDriver.StdFence
+		fence sql.Fence
 	}
 	bindings []*binding
 
 	tunnel networkTunnel.SshTunnel
 }
 
-type binding struct {
-	// Variables exclusively used by Load.
-	load struct {
-		params sqlDriver.ParametersConverter
-		keys   []string
-		insert struct {
-			sql  string
-			stmt *pgconn.StatementDescription
-		}
-		query struct {
-			sql  string
-			stmt *pgconn.StatementDescription
+func newTransactor(
+	ctx context.Context,
+	ep *sql.Endpoint,
+	fence sql.Fence,
+	bindings []sql.Table,
+) (_ pm.Transactor, err error) {
+	var d = &transactor{}
+	d.store.fence = fence
+
+	var cfg = ep.Config.(*config)
+	// Establish connections.
+	if d.load.conn, err = pgx.Connect(ctx, cfg.ToURI()); err != nil {
+		return nil, fmt.Errorf("load pgx.Connect: %w", err)
+	}
+	if d.store.conn, err = pgx.Connect(ctx, cfg.ToURI()); err != nil {
+		return nil, fmt.Errorf("store pgx.Connect: %w", err)
+	}
+
+	for _, binding := range bindings {
+		if err = d.addBinding(ctx, binding); err != nil {
+			return nil, fmt.Errorf("addBinding of %s: %w", binding.Path, err)
 		}
 	}
-	// Variables accessed by Prepare, Store, and Commit.
-	store struct {
-		insert struct {
-			sql    string
-			stmt   *pgconn.StatementDescription
-			params sqlDriver.ParametersConverter
-		}
-		update struct {
-			sql    string
-			stmt   *pgconn.StatementDescription
-			params sqlDriver.ParametersConverter
-		}
+
+	// Build a query which unions the results of each load subquery.
+	var subqueries []string
+	for _, b := range d.bindings {
+		subqueries = append(subqueries, b.loadQuerySQL)
 	}
+	d.load.unionSQL = strings.Join(subqueries, "\nUNION ALL\n") + ";"
+
+	return d, nil
 }
 
-func (t *transactor) addBinding(ctx context.Context, spec *pf.MaterializationSpec_Binding) error {
-	var err error
-	var bind = new(binding)
-	var index = len(t.bindings)
-	var target = sqlDriver.TableForMaterialization(strings.Join(spec.ResourcePath, "."), "", t.gen.IdentifierRenderer, spec)
+type binding struct {
+	target             sql.Table
+	createLoadTableSQL string
+	execLoadInsertSQL  string
+	execStoreInsertSQL string
+	execStoreUpdateSQL string
+	loadQuerySQL       string
+	prepLoadInsertSQL  string
+	prepStoreInsertSQL string
+	prepStoreUpdateSQL string
+}
 
-	// Build all SQL statements and parameter converters.
-	var keyCreateSQL string
-	keyCreateSQL, bind.load.insert.sql, bind.load.query.sql, err = buildSQL(
-		t.gen, index, target, spec.FieldSelection)
-	if err != nil {
-		return fmt.Errorf("building SQL: %w", err)
-	}
+func (t *transactor) addBinding(ctx context.Context, target sql.Table) error {
+	var b = &binding{target: target}
 
-	bind.load.keys = spec.FieldSelection.Keys
-	_, bind.load.params, err = t.gen.QueryOnPrimaryKey(target, spec.FieldSelection.Document)
-	if err != nil {
-		return fmt.Errorf("building load SQL: %w", err)
-	}
-	bind.store.insert.sql, bind.store.insert.params, err = t.gen.InsertStatement(target)
-	if err != nil {
-		return fmt.Errorf("building insert SQL: %w", err)
-	}
-	bind.store.update.sql, bind.store.update.params, err = t.gen.UpdateStatement(
-		target,
-		append(append([]string{}, spec.FieldSelection.Values...), spec.FieldSelection.Document),
-		spec.FieldSelection.Keys)
-	if err != nil {
-		return fmt.Errorf("building update SQL: %w", err)
-	}
-
-	// Create a binding-scoped temporary table for staged keys to load.
-	if _, err = t.load.conn.Exec(ctx, keyCreateSQL); err != nil {
-		return fmt.Errorf("Exec(%s): %w", keyCreateSQL, err)
-	}
-	// Prepare query statements.
-	for _, s := range []struct {
-		conn *pgx.Conn
-		name string
-		sql  string
-		stmt **pgconn.StatementDescription
+	for _, m := range []struct {
+		sql *string
+		tpl *template.Template
 	}{
-		{
-			t.load.conn,
-			fmt.Sprintf("load-insert-%d", index),
-			bind.load.insert.sql,
-			&bind.load.insert.stmt,
-		},
-		{
-			t.load.conn,
-			fmt.Sprintf("load-join-%d", index),
-			bind.load.query.sql,
-			&bind.load.query.stmt,
-		},
-		{
-			t.store.conn,
-			fmt.Sprintf("store-insert-%d", index),
-			bind.store.insert.sql,
-			&bind.store.insert.stmt,
-		},
-		{
-			t.store.conn,
-			fmt.Sprintf("store-update-%d", index),
-			bind.store.update.sql,
-			&bind.store.update.stmt,
-		},
+		{&b.createLoadTableSQL, tplCreateLoadTable},
+		{&b.execLoadInsertSQL, tplExecLoadInsert},
+		{&b.execStoreInsertSQL, tplExecStoreInsert},
+		{&b.execStoreUpdateSQL, tplExecStoreUpdate},
+		{&b.loadQuerySQL, tplLoadQuery},
+		{&b.prepLoadInsertSQL, tplPrepLoadInsert},
+		{&b.prepStoreInsertSQL, tplPrepStoreInsert},
+		{&b.prepStoreUpdateSQL, tplPrepStoreUpdate},
 	} {
-		*s.stmt, err = s.conn.Prepare(ctx, s.name, s.sql)
-		if err != nil {
-			return fmt.Errorf("conn.PrepareContext(%s): %w", s.sql, err)
+		var err error
+		if *m.sql, err = sql.RenderTableTemplate(target, m.tpl); err != nil {
+			return err
 		}
 	}
 
-	t.bindings = append(t.bindings, bind)
+	t.bindings = append(t.bindings, b)
 	return nil
 }
 
 func (d *transactor) Load(it *pm.LoadIterator, _, _ <-chan struct{}, loaded func(int, json.RawMessage) error) error {
 	var ctx = it.Context()
 
-	// Use a read-only "load" transaction, which will automatically
-	// truncate the temporary key staging tables on commit.
-	var txn, err = d.load.conn.BeginTx(ctx, pgx.TxOptions{})
-	if err != nil {
-		return fmt.Errorf("DB.BeginTx: %w", err)
-	}
-	defer txn.Rollback(ctx)
+	// We send SQL commands in batches to amortize network round-trips.
+	// All operations of Load run as a single transaction.
+	var batch = newPGBatch()
+	batch.queue(nil, "begin;")
 
-	var batch pgx.Batch
+	for _, b := range d.bindings {
+		batch.queue(nil, b.createLoadTableSQL)
+		batch.queue(nil, b.prepLoadInsertSQL)
+	}
+
 	for it.Next() {
 		var b = d.bindings[it.Binding]
 
-		converted, err := b.load.params.Convert(it.Key)
-		if err != nil {
+		if converted, err := b.target.ConvertKey(it.Key); err != nil {
 			return fmt.Errorf("converting Load key: %w", err)
+		} else {
+			batch.queueParams(nil, b.execLoadInsertSQL, converted...)
 		}
-		batch.Queue(b.load.insert.stmt.Name, converted...)
+
+		if len(batch.buf) > 1024*1024 {
+			if err := batch.roundTrip(ctx, d.load.conn.PgConn()); err != nil {
+				return err
+			}
+		}
 	}
 	if it.Err() != nil {
 		return it.Err()
 	}
 
-	var results = txn.SendBatch(ctx, &batch)
-	for i := 0; i != batch.Len(); i++ {
-		if _, err := results.Exec(); err != nil {
-			return fmt.Errorf("load at index %d: %w", i, err)
+	var loadFn = func(rr *pgconn.ResultReader) error {
+		var fields = rr.FieldDescriptions()
+		for rr.NextRow() {
+			var binding int
+			var document json.RawMessage
+
+			if err := scanRow(d.load.conn.ConnInfo(), fields, rr.Values(), &binding, &document); err != nil {
+				return fmt.Errorf("scanning Load document: %w", err)
+			} else if err = loaded(binding, json.RawMessage(document)); err != nil {
+				return err
+			}
 		}
-	}
-	if err = results.Close(); err != nil {
-		return fmt.Errorf("closing batch: %w", err)
+		return nil
 	}
 
-	// Issue a union join of the target tables and their (now staged) load keys,
-	// and send results to the |loaded| callback.
-	rows, err := txn.Query(ctx, d.load.stmt.Name)
-	if err != nil {
-		return fmt.Errorf("querying Load documents: %w", err)
-	}
-	defer rows.Close()
+	// Issue a union join of the target tables and their (now staged) load keys.
+	batch.queue(loadFn, d.load.unionSQL)
+	// We're done with our prepared statements - clean them up.
+	batch.queue(nil, "deallocate prepare all;")
+	// Commit to drop temporary load table.
+	batch.queue(nil, "commit;")
 
-	for rows.Next() {
-		var binding int
-		var document json.RawMessage
-
-		if err = rows.Scan(&binding, &document); err != nil {
-			return fmt.Errorf("scanning Load document: %w", err)
-		} else if err = loaded(binding, json.RawMessage(document)); err != nil {
-			return err
-		}
-	}
-	if err = rows.Err(); err != nil {
-		return fmt.Errorf("querying Loads: %w", err)
-	} else if err = txn.Commit(ctx); err != nil {
-		return fmt.Errorf("commiting Load transaction: %w", err)
-	}
-
-	return nil
+	return batch.roundTrip(ctx, d.load.conn.PgConn())
 }
 
 func (d *transactor) Prepare(_ context.Context, prepare pm.TransactionRequest_Prepare) (pf.DriverCheckpoint, error) {
-	d.store.fence.SetCheckpoint(prepare.FlowCheckpoint)
-	d.store.batch = new(pgx.Batch)
+	d.store.batch = newPGBatch()
+	d.store.batch.queue(nil, "begin;")
+
+	for _, b := range d.bindings {
+		d.store.batch.queue(nil, b.prepStoreInsertSQL)
+		d.store.batch.queue(nil, b.prepStoreUpdateSQL)
+	}
+
+	d.store.fence.Checkpoint = prepare.FlowCheckpoint
+	var fenceUpdate strings.Builder
+	if err := tplUpdateFence.Execute(&fenceUpdate, d.store.fence); err != nil {
+		return pf.DriverCheckpoint{}, fmt.Errorf("evaluating fence template: %w", err)
+	}
+	d.store.batch.queue(nil, fenceUpdate.String())
 
 	return pf.DriverCheckpoint{}, nil
 }
 
 func (d *transactor) Store(it *pm.StoreIterator) error {
+	var ctx = it.Context()
+
 	for it.Next() {
 		var b = d.bindings[it.Binding]
 
-		if it.Exists {
-			converted, err := b.store.update.params.Convert(
-				append(append(it.Values, it.RawJSON), it.Key...))
-			if err != nil {
-				return fmt.Errorf("converting update parameters: %w", err)
-			}
-			d.store.batch.Queue(b.store.update.stmt.Name, converted...)
+		if converted, err := b.target.ConvertAll(it.Key, it.Values, it.RawJSON); err != nil {
+			return fmt.Errorf("converting store parameters: %w", err)
+		} else if it.Exists {
+			d.store.batch.queueParams(nil, b.execStoreUpdateSQL, converted...)
 		} else {
-			converted, err := b.store.insert.params.Convert(
-				append(append(it.Key, it.Values...), it.RawJSON))
-			if err != nil {
-				return fmt.Errorf("converting insert parameters: %w", err)
+			d.store.batch.queueParams(nil, b.execStoreInsertSQL, converted...)
+		}
+
+		if len(d.store.batch.buf) > 1024*1024 {
+			if err := d.store.batch.roundTrip(ctx, d.store.conn.PgConn()); err != nil {
+				return err
 			}
-			d.store.batch.Queue(b.store.insert.stmt.Name, converted...)
 		}
 	}
 	return nil
 }
 
 func (d *transactor) Commit(ctx context.Context) error {
-	var txn, err = d.store.conn.BeginTx(ctx, pgx.TxOptions{})
-	if err != nil {
-		return fmt.Errorf("conn.BeginTx: %w", err)
-	}
-	defer txn.Rollback(ctx)
-
-	err = d.store.fence.Update(ctx,
-		func(ctx context.Context, sql string, args ...interface{}) (int64, error) {
-			// Add the update to the fence as the last statement in the batch
-			var docs = d.store.batch.Len()
-			d.store.batch.Queue(sql, args...)
-
-			var results = txn.SendBatch(ctx, d.store.batch)
-			d.store.batch = nil
-
-			for i := 0; i != docs; i++ {
-				if _, err := results.Exec(); err != nil {
-					return 0, fmt.Errorf("store at index %d: %w", i, err)
-				}
-			}
-
-			// The fence update is always the last operation in the batch
-			fenceResult, err := results.Exec()
-			if err != nil {
-				return 0, fmt.Errorf("updating flow checkpoint: %w", err)
-			} else if err = results.Close(); err != nil {
-				return 0, fmt.Errorf("results.Close(): %w", err)
-			}
-
-			return fenceResult.RowsAffected(), nil
-		})
-	if err != nil {
-		return err
-	}
-
-	if err := txn.Commit(ctx); err != nil {
-		return fmt.Errorf("committing Store transaction: %w", err)
-	}
-
-	return nil
+	d.store.batch.queue(nil, "deallocate prepare all;") // Cleanup prepared statements.
+	d.store.batch.queue(nil, "commit;")
+	return d.store.batch.roundTrip(ctx, d.store.conn.PgConn())
 }
 
 func (d *transactor) Acknowledge(context.Context) error {
@@ -456,78 +423,6 @@ func (d *transactor) Destroy() {
 	d.load.conn.Close(context.Background())
 	d.store.conn.Close(context.Background())
 }
-
-// buildSQL builds SQL statements use for PostgreSQL materializations.
-func buildSQL(gen *sqlDriver.Generator, binding int, table *sqlDriver.Table, fields pf.FieldSelection) (
-	keyCreate, keyInsert, keyJoin string, err error) {
-
-	var defs, keys, keyPH, joins []string
-	for idx, key := range fields.Keys {
-		var col = table.GetColumn(key)
-		var resolved *sqlDriver.ResolvedColumnType
-
-		if resolved, err = gen.TypeMappings.GetColumnType(col); err != nil {
-			return
-		}
-
-		// CREATE TABLE column definitions.
-		defs = append(defs,
-			fmt.Sprintf("%s %s",
-				col.Identifier,
-				resolved.SQLType,
-			),
-		)
-		// INSERT key columns.
-		keys = append(keys, col.Identifier)
-		keyPH = append(keyPH, gen.Placeholder(idx))
-
-		// JOIN constraints.
-		joins = append(joins, fmt.Sprintf("l.%s = r.%s", col.Identifier, col.Identifier))
-	}
-
-	// CREATE temporary table which stores keys to load.
-	keyCreate = fmt.Sprintf(`
-		CREATE TEMPORARY TABLE %s_%d (
-			%s
-		) ON COMMIT DELETE ROWS
-		;`,
-		tempTableName,
-		binding,
-		strings.Join(defs, ", "),
-	)
-
-	// INSERT key to load.
-	keyInsert = fmt.Sprintf(`
-		INSERT INTO %s_%d (
-			%s
-		) VALUES (
-			%s
-		);`,
-		tempTableName,
-		binding,
-		strings.Join(keys, ", "),
-		strings.Join(keyPH, ", "),
-	)
-
-	// SELECT documents included in keys to load.
-	keyJoin = fmt.Sprintf(`
-		SELECT %d, l.%s
-			FROM %s AS l
-			JOIN %s_%d AS r
-			ON %s
-		`,
-		binding,
-		table.GetColumn(fields.Document).Identifier,
-		table.Identifier,
-		tempTableName,
-		binding,
-		strings.Join(joins, " AND "),
-	)
-
-	return
-}
-
-const tempTableName = "flow_load_key_tmp"
 
 func main() {
 	boilerplate.RunMain(newPostgresDriver())
