+++ tests/materialize/materialize-postgres/cleanup.sh
