+++ tests/materialize/materialize-postgres/expected/multiple-data-types.jsonl
