+++ tests/materialize/materialize-postgres/expected/simple.jsonl
